% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CI_gamma1_evidence}
\alias{CI_gamma1_evidence}
\title{Estimate confidence interval for estimated marginal likelihood}
\usage{
CI_gamma1_evidence(temperatures, y_log_lik_t_equals_1_traces,
  confidence = 0.95, simulations = 1000)
}
\arguments{
\item{temperatures}{Numeric vector of temperatures of power posteriors. One chain will be created for each element of the vector at the corresponding temperature.}

\item{y_log_lik_t_equals_1_traces}{Numeric matrix of log probabilities of \code{y} at different temperatures (columns) in different iterations (rows).}

\item{confidence}{Numeric value of statistical confidence with which returning interval should contain the true value.}

\item{simulations}{Integer value of number of simulations to use in estimation of the confidence interval.}
}
\value{
Confidence interval as numeric vector of length 2.
}
\description{
Central limit theorem not applicable so use simulation to estimate confidence interval for evidence.
}
