% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{exact_evidence}
\alias{exact_evidence}
\title{Calculate exact evidence for model gamma = 1}
\usage{
exact_evidence(G, ...)
}
\arguments{
\item{G}{Integer matrix of variant counts per individual, one row per individual and one column per variant.}

\item{...}{Other arguments to pass to \code{\link{lower_bound_gamma1_evidence}}.}
}
\value{
Log marginal likelihood.
}
\description{
Directly evaluate evidence for model gamma = 1 by summing likelihood over all partitions of variants. Note, complexity grows exponentially with the number of variants, so is intractable for much more than ten variants.
}
\seealso{
\code{\link{gamma0_evidence}}, \code{\link{lower_bound_gamma1_evidence}}
}
