% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{expected_explained}
\alias{expected_explained}
\title{Get expected number of cases explained by pathogenic configurations of alleles}
\usage{
expected_explained(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{\link{bevimed}}.}
}
\description{
Get expected number of cases explained by pathogenic configurations of alleles
}

