% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{explaining_variants}
\alias{explaining_variants}
\title{Get expected number of variants involved in cases explained by pathogenic conigurations of alleles}
\usage{
explaining_variants(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{\link{bevimed}}.}
}
\description{
Get expected number of variants involved in cases explained by pathogenic conigurations of alleles
}

