% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{extract_conditional_prob_pathogenic}
\alias{extract_conditional_prob_pathogenic}
\title{Extract probability of pathogenicity for variant conditional on a given association model}
\usage{
extract_conditional_prob_pathogenic(x)
}
\arguments{
\item{x}{Object of class \code{x_BeviMed_m}. See function \code{\link{bevimed_m}}.}
}
\value{
Vector of probabilities of pathogenicity for individual variants.
}
\description{
Extract the probability of pathogenicity for individual variants from a \code{BeviMed_m} object.
}
\seealso{
\code{\link{conditional_prob_pathogenic}}, \code{\link{bevimed_m}}
}
