% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{extract_expected_explained}
\alias{extract_expected_explained}
\title{Extract expected number of cases explained by pathogenic conigurations of alleles from \code{BeviMed} object.}
\usage{
extract_expected_explained(x)
}
\arguments{
\item{x}{Object of class \code{BeviMed}.}
}
\description{
Extract expected number of cases explained by pathogenic conigurations of alleles from \code{BeviMed} object.
}

