% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{extract_explaining_variants}
\alias{extract_explaining_variants}
\title{Extract expected number of variants involved in cases explained by pathogenic conigurations of alleles from \code{BeviMed} object.}
\usage{
extract_explaining_variants(x)
}
\arguments{
\item{x}{Object of class \code{BeviMed}.}
}
\description{
Extract expected number of variants involved in cases explained by pathogenic conigurations of alleles from \code{BeviMed} object.
}

