% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{extract_gamma1_evidence}
\alias{extract_gamma1_evidence}
\title{Extract evidence for gamma = 1 model from \code{BeviMed} object.}
\usage{
extract_gamma1_evidence(x)
}
\arguments{
\item{x}{Object of class \code{BeviMed}.}
}
\description{
Extract evidence for gamma = 1 model from \code{BeviMed} object.
}

