% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{extract_prob_association}
\alias{extract_prob_association}
\title{Extract the posterior probability of association}
\usage{
extract_prob_association(x, by_model = FALSE)
}
\arguments{
\item{x}{Object of class \code{BeviMed}.}

\item{by_model}{Logical value determining whether to return probabilities broken down by mode of inheritance.}
}
\value{
Probability values.
}
\description{
Get posterior probability of association as numeric value, or optionally as numeric vector of length two with probabilities broken down by mode of inheritance (by passing \code{by_model=TRUE}), from a \code{BeviMed} object.
}
\seealso{
\code{\link{prob_association}}, \code{\link{bevimed}}
}
