% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gamma0_evidence}
\alias{gamma0_evidence}
\title{Calculate marginal probability of observed case-control status y under model gamma = 0}
\usage{
gamma0_evidence(y, shape = c(2, 100))
}
\arguments{
\item{y}{Logical vector of case (\code{TRUE}) control (\code{FALSE}) status.}

\item{shape}{Beta shape hyper-priors for prior on rate of case labels}
}
\description{
Marginal probability calculated exactly by integration.
}

