% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{gamma1_prob}
\alias{gamma1_prob}
\title{Calculate probability of an association}
\usage{
gamma1_prob(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link{prob_association}}}
}
\value{
Probability of association.
}
\description{
Synonym of \code{\link{prob_association}}
}

