% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log BF}
\alias{log BF}
\title{Calculate log Bayes factor comparing the variant-level model for case-control status}
\usage{
"log BF"(y, ...)
}
\arguments{
\item{y}{Logical vector of case (\code{TRUE}) control (\code{FALSE}) status.}

\item{...}{Other arguments to pass to \code{\link{bevimed}}.}
}
\value{
Log Bayes factor.
}
\description{
Calculate log Bayes factor comparing the variant-level model for case-control status
}

