% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log P_n}
\alias{log P_n}
\title{Calculate marginal probability of observed genotypes under 'null' model}
\usage{
"log P_n"(y, q_shape = c(2, 100))
}
\arguments{
\item{y}{Logical vector of case (\code{TRUE}) control (\code{FALSE}) status.}

\item{q_shape}{Beta shape hyper-priors for prior on rate of affection (i.e. being a case) amongst individuals with non-pathogenic variant combinations (i.e. they have less than \code{min_ac} variants.}
}
\description{
Marginal probability calculated exactly by integration.
}

