% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log P_v}
\alias{log P_v}
\title{Calculate marginal probability of observed genotypes under variant-level pathogenicity model}
\usage{
"log P_v"(y, G, min_ac = 1, ...)
}
\arguments{
\item{y}{Logical vector of case (\code{TRUE}) control (\code{FALSE}) status.}

\item{G}{Integer matrix of variant counts per individual, one column per individual and one row per variant.}

\item{min_ac}{Minimum allele count per individual to be considered to have a pathogenic combination of variants. '1' could correspond to a dominant inheritance hypothesis whereas '2' could correspond to a recessive inheritance hypothesis.}

\item{...}{Other arguments to pass to \code{\link{bevimed}}.}
}
\description{
Calculate marginal probability of observed genotypes under variant-level pathogenicity model
}

