% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{log_BF}
\alias{log_BF}
\title{Calculate log Bayes factor between models gamma=1 and gamma=0 given the data}
\usage{
log_BF(y, gamma0_shape = c(2, 100), ...)
}
\arguments{
\item{y}{Logical vector of case (\code{TRUE}) control (\code{FALSE}) status.}

\item{gamma0_shape}{Beta shape hyper-priors for prior on rate of case labels.}

\item{...}{Other arguments to pass to \code{\link{bevimed}}.}
}
\value{
Log Bayes factor.
}
\description{
Calculate log Bayes factor between models gamma=1 and gamma=0 given the data
}

