% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{power_posteriors_ML_sum}
\alias{power_posteriors_ML_sum}
\title{Calculate the Marginal Likelihood under v by summation over power posterior likelihood exptectances}
\usage{
power_posteriors_ML_sum(y_log_lik_t_equals_1_traces, temperatures)
}
\arguments{
\item{y_log_lik_t_equals_1_traces}{Numeric matrix of log probabilities of \code{y} at different temperatures (columns) in different iterations (rows).}

\item{temperatures}{Numeric vector of temperatures used to produce \code{y_log_lik_t_equals_1_traces}.}
}
\value{
Numeric value of estimated log marginal likelihood.
}
\description{
Calculate the Marginal Likelihood under v by summation over power posterior likelihood exptectances
}

