% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.BeviMed}
\alias{print.BeviMed}
\title{Print readable summary of \code{BeviMed} object}
\usage{
\method{print}{BeviMed}(x, ...)
}
\arguments{
\item{x}{\code{BeviMed} object.}

\item{...}{Arguments passed to \code{\link{summary.BeviMed}}}
}
\value{
Prints a summary.
}
\description{
Print summary statistics of BeviMed inference, including probability of association, probability of dominant inheritance given association and probability of pathogenicity of each variant under dominant and recessive inheritance.
}
\seealso{
\code{\link{summary.BeviMed}}
}
