% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.BeviMed_m}
\alias{print.BeviMed_m}
\title{Print \code{BeviMed_m} object}
\usage{
\method{print}{BeviMed_m}(x, ...)
}
\arguments{
\item{x}{Object of class \code{x_BeviMed_m}. See function \code{\link{bevimed_m}}.}

\item{...}{Unused arguments.}
}
\value{
Prints a summary.
}
\description{
Print summary statistics for \code{BeviMed_m} object.
}
\seealso{
\code{\link{summary.BeviMed_m}}
}
