% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{prob_pathogenic}
\alias{prob_pathogenic}
\title{Calculate variant marginal probabilities of pathogencity}
\usage{
prob_pathogenic(by_MOI = FALSE, ...)
}
\arguments{
\item{by_MOI}{Logical value determining whether to return probabilities broken down by mode of inheritance.}

\item{...}{Arguments to pass to \code{\link{bevimed}}.}
}
\value{
If \code{by_MOI} is \code{FALSE}, a vector of probabilities of pathogenicity for each variant, otherwise a matrix of probabilities of pathogenicity conditional on mode of inheritance, with a column for each variant and two rows corresponding to each mode of inheritance.
}
\description{
Calls \code{\link{bevimed}} and \code{\link{extract_prob_pathogenic}} to obtain marginal probabilities of pathogenicity.
}
\seealso{
\code{\link{extract_prob_pathogenic}}, \code{\link{bevimed}}
}

