% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stack_BeviMeds}
\alias{stack_BeviMeds}
\title{Concatenate objects of class \code{BeviMed}}
\usage{
stack_BeviMeds(...)
}
\arguments{
\item{...}{BeviMed objects}
}
\value{
\code{BeviMed} object.
}
\description{
This function could be used to stitch together consecutive chains to create one larger sampled set of states from the MCMC procedure.
}

