% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sum_ML_over_PP}
\alias{sum_ML_over_PP}
\title{Calculate marginal likelihood from power posteriors output}
\usage{
sum_ML_over_PP(y_log_lik_t_equals_1_traces, temperatures)
}
\arguments{
\item{y_log_lik_t_equals_1_traces}{Numeric matrix of log probabilities of \code{y} at different temperatures (columns) in different iterations (rows).}

\item{temperatures}{Numeric vector of temperatures used to produce \code{y_log_lik_t_equals_1_traces}.}
}
\value{
Numeric value of estimated log marginal likelihood.
}
\description{
Calculate the Marginal Likelihood by summation over power posterior likelihood exptectances
}
