% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{conditional_prob_pathogenic}
\alias{conditional_prob_pathogenic}
\title{Calculate probability of pathogencity for variants conditional on mode of inheritance.}
\usage{
conditional_prob_pathogenic(...)
}
\arguments{
\item{...}{Arguments to pass to \code{\link{bevimed_m}}.}
}
\value{
Probabilities of pathogenicity.
}
\description{
Calls \code{\link{bevimed_m}} and \code{\link{extract_conditional_prob_pathogenic}} to obtain probabilities of pathogenicity.
}
\seealso{
\code{\link{extract_conditional_prob_pathogenic}}, \code{\link{bevimed_m}}
}
