% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{expected_explained}
\alias{expected_explained}
\title{Calculate expected number of explained cases}
\usage{
expected_explained(...)
}
\arguments{
\item{...}{Arguments to pass to \code{\link{bevimed_m}}.}
}
\value{
Numeric value.
}
\description{
Use \code{\link{bevimed_m}} to perform inference under model gamma = 1 and return only the expected number of cases explained by pathogenic allele configurations.
}
\seealso{
\code{\link{bevimed_m}}, \code{\link{extract_expected_explained}}
}
