% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{extract_expected_explained}
\alias{extract_expected_explained}
\title{Extract expected number of explained cases}
\usage{
extract_expected_explained(x)
}
\arguments{
\item{x}{Object of class \code{x_BeviMed_m}. See function \code{\link{bevimed_m}}.}
}
\value{
Numeric value.
}
\description{
Extract expected number of cases explained by pathogenic configurations of alleles from \code{BeviMed_m} object.
}
\seealso{
\code{\link{expected_explained}}, \code{\link{bevimed_m}}
}
