% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{extract_explaining_variants}
\alias{extract_explaining_variants}
\title{Extract expected number of pathogenic variants in cases}
\usage{
extract_explaining_variants(x)
}
\arguments{
\item{x}{Object of class \code{x_BeviMed_m}. See function \code{\link{bevimed_m}}.}
}
\value{
Numeric value.
}
\description{
Extract expected number of variants involved in cases explained by pathogenic conigurations of alleles from \code{BeviMed_m} object.
}
\seealso{
\code{\link{explaining_variants}}, \code{\link{bevimed_m}}
}
