% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{gamma0_evidence}
\alias{gamma0_evidence}
\title{Calculate marginal probability of observed case-control status y under model gamma = 0}
\usage{
gamma0_evidence(y, tau0_shape = c(1, 1))
}
\arguments{
\item{y}{Logical vector of case (\code{TRUE}) control (\code{FALSE}) status.}

\item{tau0_shape}{Beta shape hyper-priors for prior on rate of case labels}
}
\value{
Log marginal likelihood.
}
\description{
Marginal probability calculated exactly by integration.
}
\seealso{
\code{\link{bevimed}}, \code{\link{gamma1_evidence}}
}
