% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{gamma1_evidence}
\alias{gamma1_evidence}
\title{Calculate evidence under model gamma = 1}
\usage{
gamma1_evidence(...)
}
\arguments{
\item{...}{Arguments to pass to \code{\link{bevimed_m}}.}
}
\value{
Log marginal likelihood.
}
\description{
Use \code{\link{bevimed_m}} to perform inference under model gamma = 1 and return only the log evidence/integrated likelihood.
}
\seealso{
\code{\link{bevimed_m}}, \code{\link{extract_gamma1_evidence}}
}
