% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{log_BF}
\alias{log_BF}
\title{Calculate log Bayes factor between an association model with a given mode of inheritance and model gamma = 0}
\usage{
log_BF(y, tau0_shape = c(1, 1), ...)
}
\arguments{
\item{y}{Logical vector of case (\code{TRUE}) control (\code{FALSE}) status.}

\item{tau0_shape}{Beta shape hyper-priors for prior on rate of case labels.}

\item{...}{Arguments to pass to \code{\link{bevimed_m}}.}
}
\value{
Log Bayes factor.
}
\description{
Compute log Bayes factor of an association model and model gamma = 0.
}
\seealso{
\code{\link{bevimed_m}}, \code{\link{prob_association_m}}
}
