% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.BeviMed_summary}
\alias{print.BeviMed_summary}
\title{Print readable summary of \code{BeviMed_summary} object.}
\usage{
\method{print}{BeviMed_summary}(x, print_prob_pathogenic = TRUE, ...)
}
\arguments{
\item{x}{\code{BeviMed_summary} object.}

\item{print_prob_pathogenic}{Logical value indicating whether to print list of marginal probabilities of \code{z_j = 1} for all variants \code{j} under each mode of inheritance.}

\item{...}{Unused arguments}
}
\value{
Prints a summary
}
\description{
Print summary statistics of BeviMed inference, including probability of association, probability of dominant inheritance given association and probability of pathogenicity of each variant under dominant and recessive inheritance.
}
