% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{prob_association}
\alias{prob_association}
\title{Calculate probability of association}
\usage{
prob_association(by_MOI = FALSE, ...)
}
\arguments{
\item{by_MOI}{Logical value determining whether to return probabilities broken down by mode of inheritance.}

\item{...}{Arguments to pass to \code{\link{bevimed}}.}
}
\value{
Probability of association.
}
\description{
Calculate probability of an association between case/control label and allele configuration, optionally broken down by mode of inheritance.
}
\seealso{
\code{\link{bevimed}}, \code{\link{extract_prob_association}}
}

