% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stack_BeviMeds}
\alias{stack_BeviMeds}
\title{Concatenate objects of class \code{BeviMed_raw}}
\usage{
stack_BeviMeds(objects)
}
\arguments{
\item{objects}{\code{list} of \code{BeviMed_raw} objects.}
}
\value{
\code{BeviMed} object.
}
\description{
This function could be used to stitch together consecutive chains to create one larger sampled set of states from the MCMC procedure.
}
