% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stop_chain}
\alias{stop_chain}
\title{Apply the MCMC algorithm in blocks until conditions are met}
\usage{
stop_chain(y, blocks_remaining, start_zs, start_logit_omegas,
  start_log_phis, temperatures, tolerance = 1, confidence = 0.95,
  simulations = 1000, log_evidence_threshold = -Inf,
  y_log_lik_t_equals_1_traces = matrix(ncol = length(temperatures), nrow
  = 0), full_block_traces = list(), verbose = FALSE, ...)
}
\arguments{
\item{y}{Logical vector of case (\code{TRUE}) control (\code{FALSE}) status.}

\item{blocks_remaining}{Maximum number of blocks left before termination.}

\item{start_zs}{Initial (logical) z-matrix.}

\item{start_logit_omegas}{Initial values of logit_omega (numeric vector - one value per chain).}

\item{start_log_phis}{Initial values of log_phi (numeric vector - one value per chain).}

\item{temperatures}{Numeric vector of temperatures of power posteriors. One chain will be created for each element of the vector at the corresponding temperature.}

\item{tolerance}{Maximum width for confidence_interval of log marginal likelihood to allow before stopping the chain.}

\item{confidence}{Numeric value of statistical confidence with which returning interval should contain the true value.}

\item{simulations}{Integer value of number of simulations to use in estimation of the confidence interval.}

\item{log_evidence_threshold}{Numeric value used to determine whether to stop the sampling procedure after successive blocks. If we are confident (to the level of \code{confidence}) that the evidence for model gamma = 1 is under this value, sampling is halted.}

\item{y_log_lik_t_equals_1_traces}{Numeric matrix of log probabilities of \code{y} at different temperatures (columns) in different iterations (rows).}

\item{full_block_traces}{List of outputs of calls to MCMC routine.}

\item{verbose}{To print execution progress or not.}

\item{...}{Other arguments passed to \code{\link{call_cpp}}}
}
\value{
An object of class \code{BeviMed}.
}
\description{
Sample blocks of a given size until either the estimated log marginal likelihood falls within a given confidence interval, there is sufficient confidence that the evidence model gamma = 1 is at most a certain quantity, or a certain number of blocks have been sampled.
}
