% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{summary.BeviMed}
\alias{summary.BeviMed}
\title{Create summary of \code{BeviMed} classed-object}
\usage{
\method{summary}{BeviMed}(object, ...)
}
\arguments{
\item{object}{Object of class \code{BeviMed}.}

\item{...}{Other arguments passed to \code{\link{estimate_confidence_interval}}.}
}
\value{
Object of class \code{BeviMed_summary}.
}
\description{
Create summary of \code{BeviMed} classed-object
}

