% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{summary.BeviMed}
\alias{summary.BeviMed}
\title{Create summary of \code{BeviMed} classed-object}
\usage{
\method{summary}{BeviMed}(object, gamma1_prior = 0.01,
  tau0_shape = object[["tau_shape"]], confidence = 0.95,
  simulations = 1000, ...)
}
\arguments{
\item{object}{Object of class \code{BeviMed}.}

\item{gamma1_prior}{Prior probability of that gamma = 1.}

\item{tau0_shape}{Beta shape hyper-priors for prior on rate of case labels.}

\item{confidence}{Numeric value of statistical confidence with which returning interval should contain the true value.}

\item{simulations}{Integer value of number of simulations to use in estimation of the confidence interval.}

\item{...}{Non-used arguments.}
}
\value{
Object of class \code{BeviMed_summary}.
}
\description{
Create summary of \code{BeviMed} classed-object
}

